renamed source directory to tst-hl7-pipeline added some sample messages
