+++ tst-hl7-pipeline/pom.xml
@@ -4,7 +4,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <groupId>gov.cdc.dex</groupId>
-    <artifactId>tst-hl7-testsuite</artifactId>
+    <artifactId>tst-hl7-pipeline</artifactId>
     <version>1.0.22-SNAPSHOT</version>
     <packaging>jar</packaging>
     <properties>
