+++ tst-hl7-pipeline/src/main/resources/messages/Valid-PHLIP-ORU-DataType-TX.txt
@@ -1,0 +1,9 @@
+MSH|^~\&|DMB Testing^2.16.840.1.114222.4.3.3.2.2.1^ISO|DMB Div of Lab Svcs^2.16.840.1.114222.4.1.9977^ISO|US WHO Collab LabSys^2.16.840.1.114222.4.3.3.7^ISO|CDC EPI Surv Branch^2.16.840.1.114222.4.1.10416^ISO|200708221830||ORU^R01|test0519081|T|2.3.1|||||||||PHLIP_ORU_v1.0.2^PHIN_Profile_ID^2.16.840.1.114222.4.10.3^ISO
+PID|1||DMB11664400^^^DMB Testing&2.16.840.1.114222.4.3.3.2.2.1&ISO^MR^DMB Div of Lab Svcs&2.16.840.1.114222.4.1.9977&ISO~GA16750087^^^DMB Testing&2.16.840.1.114222.4.3.3.2.2.1&ISO^DL^DMB Div of Lab Svcs&2.16.840.1.114222.4.1.9977&ISO||""^""^"^"||19730522|M||2106-3^White^2.16.840.1.113883.6.238^W^White^L~2028-9^Asian^2.16.840.1.113883.6.238^A^Asian^L|310 Bridal Path^Apt 105^Marietta^GA^30068-1234^US^H^Other Geographic^13067~PO BOX 12345^^Atlanta^GA^30342^US^M^Other Geo^13121|||||||||||2186-5^Not Hispanic or Latino^2.16.840.1.113883.6.238^N^Not Hispanic^L|||||||20070822|N
+OBR|1|EM52193^Emory Med System^2.16.840.1.114222.75.9.1.2.1^ISO|E12300^DMB Testing^2.16.840.1.114222.4.3.3.2.2.1^ISO|PLT79^Influenza Virus Detection on Isolate^2.16.840.1.114222.4.5.288|||200708190130|||ID12345^Zimmerman^John^M^MD^Dr^MD^L||||200708201530|SPT&Sputum&2.16.840.1.113883.12.70&Spu&Sputum&L^HEP&Heparin&2.16.000.0.0&H&Heparin&L^This is the method of collection^CT&Chest Tube&2.16.000.0.0&CHST&Chest&L|ID6700^Welby^Marcus^J^Jr^Dr^MD^^Emory Med System&2.16.840.1.114222.75.9.1.2.1&ISO^L|^WPN^PH^^1^678^5451200^150^voicemail~^NET^X.400^mwelby@emorymed.com|||||200708221259|||F|6584-7&Virus Culture XXX&2.16.840.1.113883.6.1&427&Viral Culture&L^1^Influenza A Virus|||PARP1234&Emory Med System&2.16.840.1.114222.75.9.1.2.1&ISO^PARF5600&DMB Testing&2.16.840.1.114222.4.3.3.2.2.1&ISO
+OBX|1|CE|FLU002^Was the patient vaccinated?^2.16.840.1.114222.4.5.232||Y^Yes^2.16.840.1.113883.12.136||||||F
+OBX|2|TX|FLU001^Was the patient receiving influenza antiviral medication?^2.16.840.1.114222.4.5.232||THIS IS A NORMAL GALL BLADDER||||||F
+OBX|3|CE|LAB514^Was this specimen outbreak related?^2.16.840.1.114222.4.5.232||Y^Yes^2.16.840.1.113883.12.136||||||F
+OBR|2|EM52193^Emory Med System^2.16.840.1.114222.75.9.1.2.1^ISO|E12300^DMB Testing^2.16.840.1.114222.4.3.3.2.2.1^ISO|PLT77^Influenza Virus Detection and Identification^2.16.840.1.114222.4.5.288|||200708190130|||ID12345^Zimmerman^John^M^MD^Dr^MD^L||||200708201530|SPT&Sputum&2.16.840.1.113883.12.70&Spu&Sputum&L^HEP&Heparin&2.16.000.0.0&H&Heparin&L^This is the method of collection^CT&Chest Tube&2.16.000.0.0&CHST&Chest&L|ID6700^Welby^Marcus^J^Jr^Dr^MD^^Emory Med System&2.16.840.1.114222.75.9.1.2.1&ISO^L|^WPN^PH^^1^678^5451200^150^voicemail~^NET^X.400^mwelby@emorymed.com|||||200708221259|||F|6584-7&Virus Culture XXX&2.16.840.1.113883.6.1&427&Viral Culture&L^1^Influenza A Virus|||PARP1234&Emory Med System&2.16.840.1.114222.75.9.1.2.1&ISO^PARF5600&DMB Testing&2.16.840.1.114222.4.3.3.2.2.1&ISO
+OBX|1|CX|LAB202^Unique Specimen ID^2.16.840.1.114222.4.5.232||DMB98011^^^DMB Testing&2.16.840.1.114222.4.3.3.2.2.1&ISO||||||F
+OBX|2|CE|40982-1^FLUBV RNA XXX Ql PCR^LN^FLUB^Influenza B RNA^L||260373001^Detected^SCT||||||F|||201501281320|||^^^09^Real Time RT-PCR^L
