+++ tst-hl7-pipeline/src/main/resources/messages/case.txt
