+++ tst-hl7-pipeline/src/main/resources/messages/lab.hl7
