+++ tst-hl7-pipeline/src/test/kotlin/gov/cdc/dex/hl7/pipeline/MainTest.kt
@@ -1,4 +1,4 @@
-package gov.cdc.dex.hl7.testsuite
+package gov.cdc.dex.hl7.pipeline
 
 import com.fasterxml.jackson.databind.JsonNode
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
