+++ tst-hl7-pipeline/src/test/resources/new-payloads/case.json
