+++ tst-hl7-pipeline/src/test/resources/verified-payloads/case.json
